Added method to read or creat new file. Slight cleaning
